+++ Sources/TuistAutomation/Utilities/BuildGraphInspector.swift
@@ -22,6 +22,11 @@ public protocol BuildGraphInspecting {
     /// Given a graph, it returns a list of buildable schemes.
     /// - Parameter graph: Dependency graph.
     func buildableSchemes(graph: Graph) -> [Scheme]
+
+    /// Given a graph, it returns a list of buildable schemes that are part of the entry node
+    /// - Parameters:
+    ///     - graph: Dependency graph
+    func buildableEntrySchemes(graph: Graph) -> [Scheme]
 }
 
 public class BuildGraphInspector: BuildGraphInspecting {
@@ -56,10 +61,18 @@ public class BuildGraphInspector: BuildGraphInspecting {
     }
 
     public func buildableSchemes(graph: Graph) -> [Scheme] {
+        graph.targets.values.flatMap {
+            $0.flatMap { $0.project.schemes }
+        }
+        .filter { $0.buildAction?.targets.isEmpty == false }
+        .sorted(by: { $0.name < $1.name })
+    }
+
+    public func buildableEntrySchemes(graph: Graph) -> [Scheme] {
         let projects = Set(graph.entryNodes.compactMap { ($0 as? TargetNode)?.project })
         return projects
             .flatMap { $0.schemes }
-            .filter { $0.buildAction?.targets.count != 0 }
+            .filter { $0.buildAction?.targets.isEmpty == false }
             .sorted(by: { $0.name < $1.name })
     }